Create User Login
This is the page displayed when you go to log in as a user to the application. The application allows new visitors of this site to become users by giving them the option of clicking on NEW USER. This allows the visitor to set a user-name and a password for them self’s. Javascript used to display a warning message if you enter wrong name or password