Create User Home Page
This is the home page for a user when the application has checked for the user name and password and they have been verified and a successful log in is completed. Supporters and players have four options on this page they can click on the buttons Hurling, Football, Squad and PlayerStatus.