Create Admin Home Page
This page is displayed when the admin has had a successful log in. The Admin has two extra buttons on their home page that the user does not. The four buttons that are in the user homepage that correspond with the four buttons in allow the admin to query the database the same as the user can on their pages. The pages are of similar layout style as the user pages however the session on these pages are different. You will need to be logged in as admin to view them through the admin side of the application 
	However the admin has a two extra buttons that allow them to update the database and be able to record statistics of matches and players. This must be allowed so that the users can few the results when they log on.