Create Start Football Match
When the "club" have a football match then the analyst must enter the fields on the the web page and the type of football match that it is (Championship, league friendly). A new session variable is created when the button is pressed  along with inputting the field information into the database.  To get to the next page of the web application then all fields must be field otherwise error message will be displayed this can be seen in the code.