Create Current Squad
When the user presses the squad button on the home page this page appears. The table shows the players in the squad of this GAA club. It shows their details such as name, position if they play hurling and football and the year they were born. The table is printed in this form by the code.Stored Procedures were used to return this information.