Create Retire A Player
The admin user can add the current squad or make one of the current members squad retired. Stored procedures are used to fill the combo box of all the players who are currently in the squad. When the submit button is pressed it displays the information of the player and if you are happy with your choice then press the retire button. Stored procedure is again used to say that this player is now retired in the database