Create Update a player
allows the admin to update a player. First a player is selected from the combo box and then their information is shown in the table. If the admin wants to change any part of this information they must fill out the fields with required information. Then press the update button, the query will not succeeded unless all fields are filled, message are displayed when the messages are empty.