Create ResultsOfHurlingGamesPlayed
When the user presses the Hurling Button in on the homepage for the user than this page will appear. There are three options on this page for the user to click on. This is done so the user can see how many games were played in the three different game types (Champions, League and Friendly's). 

When the user presses the Select for Championship button then the results of all championship games will be returned, along with the team played, the score and the year in which the match was played. A stored procedure was used to get this result when the button was pressed.