Create PlayerHurlingStatics
When the user presses the Player Status button on homepage the user has the option of picking either hurling or football. This page is displayed when the user decides on hurling. The user must select a name from the combo box and a year from the combo box. Stored procedures are used to fill these combo boxes. The user then selects the game type they would like to see the statistics for.

When a type of game button is pressed then a combination of stored procedures and MySQLi statements return the sums of all the actions that can be performed by players. The results of the sums of actions performed by the team are also returned and printed to the screen.