Create Stats During Football Game
This page is displayed when all fields have been field on "Start Football Match" page. The session variable is needed for this page because player’s statistics are being recorded on this page. The ID of the players and the action they perform must match up with an ID of a match. That is the reason for bring across the session variable, so that the information can be put into the database in the correct manner.